@@ -1,19 +1,21 @@
-# All Administrate controllers inherit from this `Admin::ApplicationController`,
-# making it the ideal place to put authentication logic or other
-# before_filters.
+# All Administrate controllers inherit from this
+# `Administrate::ApplicationController`, making it the ideal place to put
+# authentication logic or other before_actions.
 #
 # If you want to add pagination or other controller-level concerns,
 # you're free to overwrite the RESTful controller actions.
-class Admin::ApplicationController < Administrate::ApplicationController
-  before_filter :authenticate_admin
+module Admin
+  class ApplicationController < Administrate::ApplicationController
+    before_action :authenticate_admin
 
-  def authenticate_admin
-    # TODO Add authentication logic here.
-  end
+    def authenticate_admin
+      # TODO Add authentication logic here.
+    end
 
-  # Override this value to specify the number of elements to display at a time
-  # on index pages. Defaults to 20.
-  def records_per_page
-    params[:per_page] || 50
+    # Override this value to specify the number of elements to display at a time
+    # on index pages. Defaults to 20.
+    def records_per_page
+      params[:per_page] || 50
+    end
   end
 end
